tries a different design for monadic operations
It turns out that `def map[T: TypeClass](fn: T => T) = ???` isn't going
to work with parameter type inference. Even if there is a single instance
of TypeClass[T] in scope, calls like `map(x => x)` are going to fail to compile
with an error message suggesting to specify the parameter type of the lambda
explicitly.