starts implementing mergeTrees
This is a really mechanical endeavor. I wonder whether it would be useful
to implement something like `(Tree, Tree).bitraverse` with default traversal
generated by a macro. This is a significant time investment though,
so I will write the necessary code by hand first and then consider
writing a macro after I'm sure that the design is going to scale.