reworks the taxonomic API
After thinking about names for the entire day, I think that Project
was the wrong name for the abstraction of "something that has sources,
resources and dependencies".

It seems to me now that we really have three fundamentally different
(i.e. non-unifiable) use cases for this abstraction:
1) Immutable and precompiled binary artifacts
2) Immutable and not yet compiled projects
3) Mutable repl-like systems (I call them "worksheets")

1) and 2) can't be unified into one concept because you can't abstract
over the performance of scalac's compilation. 1/2) and 3) can't be unified
because in most cases one doesn't need dynamic code loading, so why expose
it for these most cases to begin with?

In any case, there's an impressive variety of use cases here, and this calls
for having a more general name than "project" for a very general concept
of something that encompasses a bunch of sources.

Towards that end, I renamed `scala.meta.projects` to `scala.meta.taxonomic`,
`Project` has become `Module` with the same API, and we now have `Domain`
which is just a list of modules. In scalahost, I'm going to experiment
with exposing the three categories of modules that were mentioned above,
and let's see what we'll be able to make of that.