removes shortcuts for the time being
I've spent quite some time trying to figure out simple ways of writing stuff like:

implicit val c1 = Universe(Artifact.Adhoc("~/Projects/foo/build/pack/foo.jar", "~Projects/foo/src"))
implicit val c2 = Universe(Project.Sbt("scalameta", "~/Projects/core"))

It was really hard, because with all the capabilities for library design
that Scala provides, one can apply a lot of tricks to make things look
more concise.

And then, when I noticed that I'm already knee-deep in this bog,
I decided that it's just not worth it. The lines above aren't actually
that bad, and what's important - they make stuff explicit, which avoids
all the ambiguities of Scalahost.mkProjectContext("...", "..."), and that
was the goal for this design exercise in the first place.

The only shortcut that I'm really fine with is `"orgId" % "artId" % "version"`
to specify a maven artifact. This is well-established and obviously unambiguous,
so I'm letting it live.