Module.dependencies don't belong here
While programming a bit more on paper, I figured out another interesting thing.

With the API that existed before this commit, one could write something like
`("com.chuusai" %% "shapeless" % "2.2.5").dependencies`, but that would
require having a taxonomic context in scope.

So, in essence, if one wanted to figure out the dependencies of some module,
they would have to wrap this module in a universe and then take that module again
and call .dependencies on it.

This looked rather stupid!

I tried to make this scenario better, without particular success, but then
realized that we actually have zero immediate need in Module.dependencies.
Module.sources (and, by extension, Module.resources) are reasonable, because
that's essentially one of the main entry points into the scala.meta API,
but figuring out dependencies of modules has never been a promise of scala.meta.

Maybe in the future, we'll add something like that, but for now I'm just removing it.