I was totally wrong about Module.dependencies
It looks like I've been conflating taxonomic and semantic contexts in one entity,
whereas in fact they are completely separate.

The former simply loads trees from TASTY or whatever optimization
we can come up with for particular types of modules, but it doesn't need
to know what the loaded trees mean.

The latter knows how to make sense of denotations and typings given
a typechecker and a collection of sources, but it doesn't need to know
how these sources were loaded.

This, for instance, means that you don't need to create `Universe(artifact)`
in order to figure out what this artifact depends on. Also, TaxonomicContext.domain
doesn't make any sense, because an artifact resolver or an sbt interface
doesn't actually care about other artifacts/projects that happen to exist
somewhere nearby.

This also explains the troubles that I've been having with semantic APIs
like .defn returning emulated versions of definitions instead of full-fledged sources.
At the moment, scalahost conflates taxonomy and semantics, which makes it
easy to make mistakes in the implementation.

As a result, I've reinstated Module.dependencies, removed Module.domain,
and nuked Worksheet.append (this one I have no idea what to do with).
I'm also feeling kinda sad now, because it's back to the drawing board again.