new day => new vision
Here's a refinement of prior attempts to strike balance between semantic
and taxonomic aspects of scala.meta. Let's hope it works:

1) Semantic and taxonomic contexts are orthogonal to each other in the sense
that in the general case neither needs another one to do their job.

In practice, a semantic context might need a taxonomic context to properly initialize
the underlying compiler instance. Also, depending on the platform, a taxonomic context
might need a semantic context to load TASTY from platform-specific metadata,
but let's hope it doesn't come to this, because it'll create nasty cyclic dependencies.

Think of a semantic context as of a compiler, and of a taxonomic context
as of a maven/ivy resolver (the taxonomic context is more general than that,
but the analogy still stands).

Now, if you ponder this, a compiler and a resolver should definitely be
two different things, so that we don't conflate matters unnecessarily.
Also, a compiler doesn't need to know about any resolvers, because it
might accept dependencies just as paths to class directories and jars.
However, it might want to use a resolver to provide a more convenient
interface to the users (so that they can specify dependencies as maven ids
rather than download the artifacts, manage the cache and provide links to
cache entries).

2) Semantic contexts now have `def domain: Domain`. This was the piece
that was obviously missing, but I couldn't quite put a finger on it.
Now, when we have an abstraction for a domain, it all makes perfect sense.

3) Projects are gone, because they ended up being too hard. The notion of
"maybe compiled" has insane complexity - both theoretically (because it
introduces mutability which we have successfully avoided until now) and
practically (because it would require to programmatically interface with
build tools).

The original motivation behind creating projects - support for GlobalContext
via Project.Adhoc - is now served very well by simple, immutable Module.Adhoc,
without any need for additional complexity.

4) Worksheets are gone as well, because again, they conflated
the immutable notion of taxonomy and the mutable notion of interactive compilation.
After I created a dedicated abstraction for that: interactive.Context,
things have become much better. This has satisfied the immediate need
in loading compilation units dynamically, without getting us into
rabbit holes, plus, created the foundation for the potential future
meta-based presentation compiler API.