Taxonomy now also takes a dialect
Makes sense, if you ask me, because in order to parse source files
(which is an essential operation for the standard taxonomy, as it needs
to correlate syntactic and semantic sources) you need a dialect.

Requiring any kind of parsing and prettyprinting to take an implicit dialect
is really useful in exposing and fixing logical errors!