Module => Artifact
On a second thought, the word "module" is too platform-specific.
Is it a jigsaw module? Is it a scala.reflect module? Confusing.

I'm not completely satisfied with "artifact", because in my mind
"artifact" stands for something already compiled, and Artifact.Adhoc
doesn't fall into this category, because its sources can be unattributed.

However, I don't want to spend any more time on modelling and/or naming
the taxonomic part of our API, so "artifact" is going to suffice.