works around a craziness in BatchSourceFile
scalac never ceases to amaze me.

Really, if you're not compelled to expose your inner workings to the outer world
(for the compiler, this is equal to having a public metaprogramming API),
then all kinds of ugly hack may creep in in unexpected places.

I mean, if you're having problems with positions in your parser, why not
go to your slurping layer and postprocess all incoming sources so that they end in newlines?
What can possibly go wrong???