delays global.Run creation
Creating a global run forces lazy computation of the compiler's classpath.
This, in turn, prevents us from modifying the classpath.

Proxy.indexCompilationUnits relies on the fact that it can take
the initial domain, flatmap its binaries and update the compiler's classpath
with that information in order to include the initial domain into the symbol table.

I tried using global.extendCompilerClassPath to perform a hot update
of the symbol table, but it ended up being really picky, so we really
need to initialize the compiler at the right moment of time.