introduces semantic.Environment
It would seem that the problems that early users have with eager naming
will force us to at least introduce the fundamental data structures for
proper hygiene. This is the first step.

Which problems? Here we go:

1) Eager naming is an unprincipled hack, so it only works for few select
particular cases. No idea how to scale it further.

2) Eager naming relies on c.typecheck, and that one sometimes leads to
spurious cyclic references (see #174).

3) Eager naming requires clients to have respective definitions on their
classpath. Sometimes this is impossible, and then it becomes a dealbreaker.