exterminates the current approximation of hygiene
As described in https://github.com/scalameta/scalameta/issues/156,
the current implementation is an evolutionary dead-end, so we have to
remove it and the earlier the better.

My only concern is that we're losing the ability to conveniently get
trees with pre-populated denotations, but I think we'll have to live
with that.