introduces Context.typecheck
With the semantic model finally in place, and .denot, .typing and .expansion
having proven to be useful for modelling semantic information, we can
finally remove the flair of mystery behind Term.desugar, Term.tpe and
Term.Param.tpe, unifying all them behind Context.typecheck.

This also fully empowers contexts to make sense of arbitrary user input.
Previously, we relied on eager attribution in quasiquotes (or on tree
conversion, but that's not always available) to provide adequate input trees
for context APIs. Now, we can just fabricate any input tree without
worrying about its denots and pass it to a context.