redoes tree equality
I've finally given up on having Tree.== working hygienically, because:
* Such comparisons can't take implicit arguments
* Which means that comparisons can't do any semantic operations
* Which means that names have to be preresolved
* Which means that quasiquotes must do eager naming
* And that is totally impossible in general case

Previously it was sort of working, because I believed that we can approximate
eager naming well enough, but then I realized (thanks to @sschaef)
that sometimes it's just impossible, because quasiquotes may refer to something
that will only become known at runtime.

So, we really have to bite the bullet, have Tree.equals and Tree.hashCode
to express referential equality and forget about having out-of-the-box
Maps and Sets support trees.

On the plus side:

1) We can easily have multiple equalities now (e.g. this commit introduces
=== for semantic equality and =:= for semantic equivalence).

2) We can have equalities take an implicit semantic context and typecheck
their arguments, failing with an exception if the user has made a typo
(instead of stupidly returning false, like we did before).

3) We can strongly type equalities, avoiding obvious nonsense,
e.g. an accidental comparison between a Ref and a Member.

4) We get to join the cool kids and whine how bad universal equality is
and how it impedes our self expression.