implements AllowEquality
The initial idea is to statically disallow comparing unrelated subclasses of Tree,
where "unrelated" means those that don't have any common parents except for
Tree, Stat and Scope. Equal types (e.g. Tree and Tree) can still be compared,
it's just that if one of the comparees is more precise, it usually indicates an error.

This is an experiment. The heuristic seems useful enough to avoid an entire
class of errors, e.g. `ref.defn == t"SomeClass"`, but still not arbitrary enough
to be predictable and not annoying. Let's see how this works in practice.