introduces Typing.Recursive
This is something that's very obvious and that has been completely overlooked.

If you have `object Foo`, then the typing of Term.Name("foo") is going to be
Type.Singleton(Term.Name("foo")), and the typing of that Term.Name("foo")
is going to be Type.Singleton(Term.Name("foo")), and so on, and so on.

In order for this not to end in a stack overflow, we need to tie the knot somewhere.
It looks like the best place of doing that is the Typing adt.

Notice how we faced roughly the same problem with Expansion, having to
introduce Expansion.Identity to also accommodate recursion.

Typing.Recursive is extremely trivial at the moment. It doesn't have any
parameters at all, and just its mere presence indicates that we're looking
at a singleton. I don't think that any other terms in Scala can have types
that refer to those terms (or parts thereof), so let's keep this as simple as possible.

In this commit, I also had to fix show[Semantics], because I realized
that it no longer deduplicates entries.