introduces Type.Method
This is actually a big conceptual deal, because it breaks the idea
of only using existing syntax to metaprogram Scala. We sort of already
violated the principle with Type.Lambda, but there I had excuses, and now
I can't come up with any.

Anyway, the alternative is to have `Term.tpe` calls on method references
to crash, effectively requiring all calls to `Term.tpe` to be wrapped
in a try/catch, which goes against another principle of scala.meta that
fatal errors shouldn't need to be handled everywhere. What's especially
annoying in such a hypothetical design is that even when analyzing TASTY,
which is guaranteed to be well-typed (hence guaranteed to have no fatal errors
when doing semantic APIs) - a very big use case of utilizing scala.meta -
exception handling would be required.

Also changes the syntax of Type.Lambda, so that a printout of
Type.Lambda(..., Type.Method(...)) looks exactly like the relevant part
of a method declaration.