less stressed out logic in Type.ctorRef
`tpe.members(ctor.defn).tpe` was put there as a safeguard to make sure
that we don't mess up prefixes. This ended up being a problem for the converter,
because it calls ctorRef, that calls into the semantic context (which is
the converter itself), and that creates nasty recursions.

What I realized though is that we could just require our users to specify
the correct prefix in ctor.typing, and then we avoid this problem altogether.