classifies and refactors internal tree APIs
Classifies all non-public tree APIs into internal ones (usable by hosts)
and private ones (usable only by the scala.meta framework).

I would like to distinguish these categories by access modifiers, but
that's not easy: 1) private[meta] doesn't work, because hosts are
defined in scala.meta.something, 2) protected almost works, because
almost all private APIs are protected and are used only within definitions
of AST classes, but for the former internalCopy (now privateCopy) this
isn't true. Therefore, I've renamed all private methods to begin with "private".

This commit also removes withDenot and withTyping, merging them into withAttrs.
withEnv and withExpansion have survived, because they have been found to
be fundamentally necessary for the framework.

The new vision of scala.meta tree internals is described in the update
to hosts.md. Now, when this new design finally makes sense to me, I can start to
actually implement it. Hope it will survive an intense contact with the real world.

@densh has also suggested doing away with Tree.typing and Tree.expansion,
caching them not inside the trees, but inside hosts that produce them.
This might be a good idea to reduce the complexity of tree internals,
because it's quite challenging at the moment with 4 internal states and
more than a handful of internal fields. However, I'm a bit scared of doing
such a drastic refactoring within the time constraints of an imminent
conference, so I'm leaving this for future work.