removes replIntegration from the build
Much like in ReplSuite, I'd like to create a mirror that is built from
a path to scala-library.jar. I'd make `initialCommands in console` depend
on `fullClasspath in console`, but that's illegal, because initialCommands
is a setting and fullClasspath is a task. I tried several other approaches,
but all of them were unsuccessful, so I decided to simply remove this functionality.