introduces support for sensitive logging
Instead of `if (Debug.xxx) ...`, we now have a higher-level `Debug.logXxx(...)`.
This has allowed us to implement something cute.

Since some of our logs are quite invasive and can cause untimely forcing
of typings, we need a facility to detect points of time when logging is not
welcome and to log stuff at appropriate moments.

This has been implemented by introducing `Debug.delayingSensitiveLogs`
that automatically delays thunks passed to `Debug.sensitiveLog` (and that
is, in turn, can be called by some of the `Debug.logXxx` methods) and then
dumps them in batch when the enclosed block of code is exited.