temporarily removes TQL
TQL is an amazing prototype of an alternative, FP-style API to tree
traversal and transformation.

There are open problems (underwhelming performance and complex API surface),
but if they are solved, it has a good chance to supplant the more traditional
Traverser/Transformer pattern.

Unfortunately, features specific to TQL haven't gained much traction
among existing users of scala.meta, and TQL's shortcomings are real.
Therefore, I'm going to exclude it from the upcoming 0.1 release,
but I hope to resurrect it in the future.