merges foundation, classifiers and prettyprinters
I'm not sure that this fine-grained distinction was actually useful.

First of all, if we follow the logic "anything that's common and shows up
in the public API - goes into a separate project", then we should've created
a project for Convert.

Secondly, three helper projects (four with Convert) is definitely too much.
The complexity really shows in modules.svg.