Tokens is no longer Input
Generalizing Array[Char] inputs to Seq[Token] inputs was a nice idea,
theoretically, but in practice it: A) was never used except for the internals
of the quasiquoting macro, B) led to weird effects.

To expand on the last point, here are the problematic consequences of
generalizing the inputs:
  1) If an AST is assembled from a heterogeneous bunch of tokens,
     then it's unclear what's its position, because it can span tokens
     from multiple contents.
  2) As a result, it's unclear how we do error reporting under such conditions.
     In the quasiquoting macro, error positioning always makes sense,
     but in the general case it sometimes doesn't.

Therefore, I'm undoing the generalization and specialcasing quasiquote
parsing in the quasiquoting macro.