introduces @none
@none annotation is an analogue of @leaf that expresses one additional fact:
that the underlying object is a null object of the corresponding ADT family.

I didn't want to do anything magical, e.g. autogenerating the isEmpty/etc
methods to the family if one of the family members is @none. Therefore,
I'm also introducing the new Optional trait that captures the essence
of this logic.