demotes Tree.tokens
Tokens are now calculated from Tree.pos (if the tree has a non-empty pos)
and Tree.syntax (if the tree has an empty pos), not vice versa.

Tree.tokens is now correctly taking a dialect. This was a major oversight
in the previous implementation, which gave rise to #220.

Tree.syntax (and other syntax methods) are now correctly taking a dialect,
which provides the first part of the fix to #220.

Tree prettyprinting is no longer based on tokens. This provides a fix to #171
and the second part of the fix of #220. Unfortunately, this also
brings us to square zero with respect to #149. Maybe #226 will provide
some help here, maybe not, but in any case we needed to replace the broken
implementation of prettyprinting.

Quasiquote parsing is also no longer based on tokens. Assembling a token stream
from string parts and intermediate unquotes was cute (especially in comparison
with what we did in scala.reflect), but it brought too many complications.
The most prominent of them was the notion of Tokens <: Input, which had
caused shotgun changes all over the framework.