removes duplication of instance and extension methods
Back then, there was an idea that we should take some core extension methods
are copy/paste them onto core classes of our data model. This improves
the tooling story and arguably improves discoverability.

Unfortunately, this also leads to code duplication. Even if we share
implementations (by just redirecting instance methods to extension methods),
there is duplication of signatures. We had a hypothesis that this wouldn't
be significant, however a recent refactoring experience provided evidence
against this hypothesis.

When I tried to replace class Tokens with type Tokens = Seq[Token],
I had to do something with Tokens.syntax and Tokens.structure.
If I just removed them, that'd be inconsistent with the "core classes"
part of the principle explained above. If I didn't remove them, I'd have
to find new home for them (because type aliases don't have companions),
and in that case there would be problems with implicit priority.

/cc @olafurpg