removes Position.point
I'm convinced by @densh's feedback about not pretending that we support
Position.point, when in fact we don't. This will proliferate the design
pattern that comes from scalac and that may very well originate in some
of its limitations that we've already overcome.

Being potentially incompatible with scalac wrt error messages is a valid concern,
but let's cross that bridge when we get there. We don't know what a potential
solution would look like, so leaving a stub to possibly support it
is not very prudent.