correct error message for ...$ for Term.ApplyInfix, Pat.Extract and Pat.ExtractInfix
These AST nodes only support Seq[Arg], not Seq[Seq[Arg]],
because this is what Scala grammar allows.

Here's a funny gotcha that follows from this:
scala> 16:32 ~$ parse "x + (y)(z)"
[[syntax trees at end of parser]]// Scala source: tmpqW0GnW
x.$plus(y(z))

Anyway, jokes aside, this means that we need to disallow ...$
in argument position for these nodes.

Term.ApplyInfix and Pat.ExtractInfix are created by an infrastructure
that's based on argumentExprs(). argumentExprs() is also called to
parse arguments for regular calls, so this means that we have to validate
the absense of ...$ elsewhere.

Pat.Extract is also based on a very generic infrastructure, so we specialcase
that one too.