attempt to fix https://github.com/scalameta/paradise/issues/10
typeOf/symbolOf/hygienicRef for existential types have been replaced
with equivalent constructs that don't involve reification. To see why
that's necessary, refer to the aforementioned issue.

I don't like the fact that I can't test this in an automated fashion
(well, at least I can't come up with a way of doing that at the moment).