work around #448
We can't fully fix #448 at the moment because of the syntactic ambiguity
described in #223.

When the user writes q"class C extends $foo", we can't really know whether
they meant to insert/extract a constructor or a template.

Therefore, what we expect people to do now is to write q"class C extends $foo {}",
in which case they unambiguously point out that they want to work with a parent.
It turns out that this didn't work before, and now I fixed that.
I also improved the error message to clearly explain what needs to be done.

I hope that in the future we'll be able to do better here.