Properly tokenize xml literals, fixes #356.
This commit makes xml literals tokenize like string interpolators.
The solution works roughly like this:

In LegacyScanner:
1) when encountering an xml literal, run a fastparse xml parser to get
   the positions of scala expressions inside the literals.  Populate a
   `xmlLits: Map[Int, Int]` with start positions of the scala
   expressions as keys and their respective end positions as values.
2) in nextToken(), if the current offset has a key in `xmlLits`.
   return an XMLLIT LegacyTokenData and scan until the end of the
   xml literal part. Otherwise, proceed as normal.

In ScalametaTokenizer: we use a similar approach as with interpolated
strings.  When encountering an XMLLIT LegacyTokenData, we emit a
Token.Xml.Part and then run ScalametaTokenizer.loop on the embedded
scala expression up until until the end curly brace.  Then we repeat.