Bump sbt to 0.13.12
This was because I know of a fix the the incremental compiler for
package objects, which scala.meta uses a lot.

Sadly, this doesn't improve recompilation like I expected because the
incremental compiler bails at the sight of macros, which scala.meta also
uses a lot.

But it's still a good idea to upgrade in my opinion.