Fix #470: Make sure that `has` and `getAll` work
The previous implementation of `getAll` didn't work as expected.

In short, it checked that the runtime classes are the same, when they are not: one represented the interface of a modifier, whereas the other one represented the actual implementation, that had a different class name (generated by a macro) and was a subclass of the first one.

The following fix makes sure that tags are appropriately comparable by relying on the existing Mod classifiers.