Don't accept `final trait A` + add final test suite
According to scalac, final trait `A` should not parse because traits are
always `Abstract`. This changeset modifies the parser to add the
`Abstract` flag in every trait and adds utils to throw syntax errors for
illegal combination of modifiers.

`SuccessSuite` is modified because it was parsing `final trait`, which
is rejected by the scalac's parser.