Fix #46: Enable comments inside quasiquotes
Quasiquotes preserve now their parsed origin and `tokens` allow you
to access any comments from the original source.

The fix consists in tokenizing, getting the token positions, creating
the origin `Origin.Parsed` and setting it to term quasiquotes. Pattern
quasiquotes cannot benefit from this, because comments are only
preserved in the token level, not the tree level. Therefore, one cannot
pattern match on a quasiquote with a comment (because of `unapply`).

There are some tests failing because `show[Structure]`, implemented
in `TreeStructure.scala`, does not handle `Unquote`. Therefore, it
outputs an empty string when the spliced quasiquote variable should be.
Fixing this requires changing the accessibility of Unquote and
"magically" detecting the value of that spliced variable. The actual
meta tree has all this information, but I find it difficult to map it to
the tokens. Not a trivial task.