revive scalahost
This commit moves the converter part of scalameta/paradise to our old friend
scalameta/scalahost.

Several projects, including scalafix, are in need of the converter without
the additional burden of the entire paradise, so I’m creating a dedicated
library for that.

Moreover, this will be the place to implement our upcoming semantic API
courtesy of early experiments in scalafix and nirvana.