change versioning scheme
We no longer use PR numbers to derive prerelease versions, because
such versions aren’t guaranteed to be ordered lexicographically.
Earlier pull requests can be merged later, and vice versa,
which brings significant confusion.

The new versioning scheme is very similar to git-describe.
A typical prerelease version looks like 1.6.0-344-a9f6f83e,
where the first number after the dash is the number of commits between
HEAD and v1.0.0, and the second number is the sha of HEAD.
This is a little bit more verbose than the old scheme, which only featured
one number, but that’s what it takes to be (almost) lexicographic and unique.

Finally, if an artifact is built from a dirty repository, i.e.
if the sources aren’t uniquely identified by the current sha, then we
append a timestamp, e.g. 1.6.0-344-a9f6f83e.1487281742717.