Use protobuf to persist semanticdb.
The current persistence scheme for the semanticdb relies on an ad-hoc
serialization format. This is problematic if we want to add/remove fields
while keeping compatibility with existing semanticdb.

This commit migrates the serialization format to protobuf, which will
make it easier to evolve the schema without breaking compatibility
with existing semanticdb. Moreover, by delegating the serialization task
to an established library, we avoid the implementation cost of
implementing custom serializers and de-serializers.

Why protobuf instead of thrift+scrooge?

1. scalapb supports Scala.js, which may come in handy when
   building web-based tools such as code-search/refactoring UIs.
2. I personally have more experience with protobuf.

Why protobuf instead of scodec?

1. scodec is Scala only, while protobuf has bindings for a lot of
   languages. Someone might want to build tools/applications on top
   of semanticdb using another language than Scala.
2. protobuf has more documentation and is more widely used.
3. I don't have much experience with scodec to judge how hard it would
   be to design a protocoal that makes it easy to add/remove fields from
   the schema without breaking backward compat.

If it turns out that thrift is a better fit for this use-case, I think
it will still be possible to migrate. The semanticdb will only contain
simple field types to start with (mostly int32/string), which are
supported by both protobuf and thrift.