Restrict secrets to images.
Drone v0.5 changed secrets so that they can be restricted to images.
See https://github.com/drone/drone/issues/1727. This is a great
feature since it makes it possible to expose only secrets to use
different docker images for tests and publish and expose only secrets
in the publish step.