rename Database.symbols to Database.names
I was about to add the symbol table to the database, and that's when I realized
that it's the part of the database that should really be called `symbols`.

While freeing up the field name, I realized that we should call
the existing field `names`. Incidentally, this is also consistent with
name shortening that we use for `messages: Seq[CompilerMessage]`.