remove support for unattributed trees in Mirror
In the spirit of continuous simplification of our semantic infrastructure,
I’m removing support for unattributed trees in both offline and online
mirrors.

******* Background *******

From the early days of reflect macros (2011/2012), I’ve been dreaming
about a magical API that would blur the distinction between
unattributed and attributed trees and automatically convert between one another.
In the world of complete craziness of unabstracted compiler internals,
this dream gave me a reason to live.

In 2013, Jason and I made a stab at solving the problem within scala.reflect.
We came up with a promising technique and an interesting research direction,
but neither of us had time to follow that through.

In 2014, Denys and I founded scala.meta where we proclaimed that
metaprogramming APIs should be user-friendly and, therefore, all trees
are created equal, and now it’s our problem to figure out how to implement
that. In our manifesto, we also published some really serious wishful thinking,
and the adrenaline rush from that told us that everything’s gonna be allright.

When the time came to actually implement what’s been promised
(in the beginning of 2015), it turned out that unification is harder
than I thought. The looming deadline didn’t leave me much time to think,
so I fell back to runtime typechecking of converted reflect trees and
moved on. Since we didn’t support conversion of attributes, this effectively
amounted to supporting either fully attributed or fully unattributed trees.

Time went on, and in the big cleanup of 2016, the prototype of the scala.meta
semantic API got removed completely. I felt a lot of relief, as I didn’t
have to deal with the problem of attributed/unattributed trees anymore.

Nonetheless, in 2017, the semantic API got reintroduced and with it
the problem resurfaced once again. And again, I hacked up a stub that
involved runtime typechecking and moved on to implement more important
functionality.

Last week, I realized that enough is enough. Having two distinct concepts
of unattributed and attributed trees - either at the API level or at the
implementation level - is simply too hard. Life is too short for this stuff,
so I give up and move on. There, I said it.

******* Impact *******

On a purely technical note, the decision to remove support for unattributed
trees in the semantic API significantly simplifies our codebase.

Offline mirrors no longer need to instantiate a global, and as a result
no longer need to take a plugin path. This greatly simplifies user experience
and actually opens up new avenues, because we no longer need to carry
compilers around.

Another pleasant fact is that we can remove Address.Snippet
(whose sole purpose of existence was to accommodate runtime typechecking)
and with it the whole Address data structure.

On the negative side, we can no longer use quasiquotes to conveniently
obtain symbols (as in `q”Int”.symbol`). However, that didn’t work well anyway,
because for this technique to operate robustly, one actually has to provide
fully qualified names, which kinda defeats the purpose of using quasiquotes
in the first place. Therefore, I think it’d be better to completely get rid of
this wishful thinking and create a new API from scratch.