reshuffle some files between semantic, scalahost and scalahostNsc
1) OfflineMirror along with the factory methods that create its instances
got moved to semantic, according to the new vision of OfflineMirror
not depending on a compiler.

2) semantic now depends on parsers, because OfflineMirror uses .parse[Source]
in order to convert sourcepaths to sources. That’s a valuable insight.

3) ScalahostReporter got moved to scalahost. The only remaining logic
in scalahostNsc is the plugin definitions and hijacking incantations.

4) Exposed all data structures inside scala.meta.semantic.v1 in the public API.
Back then, I wanted to pretend that the API surface is really small,
including just Symbol and Database, but the reality is different,
so it’s time to embrace it.