remove the distinction between semantic.v1 and semantic
This is yet another victim of my recent search-and-destroy-wishful-thinking spree.

Let’s face the facts - the “v1” semantic API based on a “temporary”
and “low-level” concept of a symbol has been wildly successful, becoming
the foundation for next-generation tools from Scala Center and Twitter
even in its current prototypical state.

Compare this with the story of the unicorn vision of the semantic API
designed in 2013. That vision sounded quite good on paper, but it has never
been implemented, even as a prototype, because some pesky implementation
detail has always got in the way.

The only logical conclusion from comparing these two stories is that
the unicorn API doesn’t work, and that the “temporary” API was the way
to go all along.

In order to acknowledge that, I’m throwing away the v1 prefix to the
“temporary” API to make it the one and only approach that we’re investing in.